Reuse mapping validation cases across generated outputs
