+++ tests/data/expected/main/jsonschema/conditional_json_equality/enum_True.py
@@ -3,6 +3,7 @@
 
 from __future__ import annotations
 
+from collections.abc import Mapping as _Mapping
 from typing import Any, ClassVar
 
 from pydantic import BaseModel, Field, model_validator
@@ -19,7 +20,7 @@ class _JsonSchemaRuntimeValidationBase(BaseModel):
 
     @classmethod
     def _validate_json_schema_conditional_required(cls, data: Any) -> Any:
-        if not isinstance(data, dict):
+        if not (isinstance(data, dict) or isinstance(data, _Mapping)):
             return data
         for rule in cls.__json_schema_conditional_required__:
             condition_matches = all(
